Super-Resolution-main/be/sr-server/requirements.txt
fastapi
pydantic
celery
python-dotenv
supabase
uvicorn
PyJWT
python-dotenv
python-multipart
colorlog
uuid7
reportlab 
httpx[http2]
requests
